+++ config/fcos/v1_5_exp/translate.go
@@ -68,17 +68,34 @@ func (c Config) ToIgn3_4Unvalidated(options common.TranslateOptions) (types.Conf
 		return types.Config{}, translate.TranslationSet{}, r
 	}
 	r.Merge(c.processBootDevice(&ret, &ts, options))
+
 	for i, disk := range ret.Storage.Disks {
-		// In the boot_device.mirror case, nothing specifies partition numbers
-		// so match existing partitions only when `wipeTable` is false
-		if !util.IsTrue(disk.WipeTable) {
-			for j, partition := range disk.Partitions {
-				// check for reserved partlabels
-				if partition.Label != nil {
-					if (*partition.Label == "BIOS-BOOT" && partition.Number != 1) || (*partition.Label == "PowerPC-PReP-boot" && partition.Number != 1) || (*partition.Label == "EFI-SYSTEM" && partition.Number != 2) || (*partition.Label == "boot" && partition.Number != 3) || (*partition.Label == "root" && partition.Number != 4) {
-						r.AddOnWarn(path.New("json", "storage", "disks", i, "partitions", j, "label"), common.ErrWrongPartitionNumber)
+		for p, partition := range disk.Partitions {
+			//if the partition is not nil and is root
+			if partition.Label != nil {
+				if *partition.Label == "root" {
+					if partition.SizeMiB == nil || *partition.SizeMiB == 0 {
+						for _, ap := range disk.Partitions {
+							if ap.Number == partition.Number+1 {
+								if ap.StartMiB == nil || *ap.StartMiB == 0 {
+									r.AddOnWarn(path.New("json", "storage", "disks", i, "partitions", p, "number"), common.ErrRootConstrained)
+								}
+							}
+						}
+					} else if *partition.SizeMiB < 8192 {
+						r.AddOnWarn(path.New("json", "storage", "disks", i, "partitions", p, "size_mib"), common.ErrRootTooSmall)
 					}
 				}
+
+				// In the boot_device.mirror case, nothing specifies partition numbers
+				// so match existing partitions only when `wipeTable` is false
+				if !util.IsTrue(disk.WipeTable) {
+					// check for reseved partlabels
+					if (*partition.Label == "BIOS-BOOT" && partition.Number != 1) || (*partition.Label == "PowerPC-PReP-boot" && partition.Number != 1) || (*partition.Label == "EFI-SYSTEM" && partition.Number != 2) || (*partition.Label == "boot" && partition.Number != 3) || (*partition.Label == "root" && partition.Number != 4) {
+						r.AddOnWarn(path.New("json", "storage", "disks", i, "partitions", p, "label"), common.ErrWrongPartitionNumber)
+					}
+				}
+
 			}
 		}
 	}